extract logic to abstract class, make ContextAwareQueryRewriter an interface again, do not save member variable in abstract class #181
